app/main.py
from fastapi import FastAPI
import yaml

app = FastAPI()

# Load input OpenAPI specification
def load_openapi_spec():
    with open('openapi3_1.yml', 'r') as file:
        return yaml.safe_load(file)

# Example root endpoint
@app.get("/")
def read_root():
    return {"message": "Hello, World!"}

# Additional routes can be added here
