requirements.txt
fastapi
uvicorn
typesense
pydantic
PyYAML

